Deduplicate fresh_item_substs
